playing around with stuff

+++ torchserve/project_action/project_action_handler.py
@@ -1,12 +1,12 @@
 from ts.torch_handler.base_handler import BaseHandler
 import torch
-from sentence_vectorizer import vectorize_sentence
+#from sentence_vectorizer import vectorize_sentence
 
 
 class ModelHandler(BaseHandler):
   def preprocess(self, data):
 	  #sentence = data.get("data")
-  	return torch.zeros(1,10,300)
+  	return torch.zeros(1,10,300).float()
 
   def postprocess(self, model_pred):
     index2label = {0: 'Create',1: 'On Target',2: 'At Risk',3: 'Danger',4: 'Completed'}
